+++ js/source/image_source.js
@@ -2,7 +2,7 @@
 
 var util = require('../util/util');
 var Tile = require('./tile');
-var LatLng = require('../geo/lat_lng');
+var LngLat = require('../geo/lng_lat');
 var Point = require('point-geometry');
 var Evented = require('../util/evented');
 var ajax = require('../util/ajax');
@@ -14,16 +14,16 @@ module.exports = ImageSource;
  * @class ImageSource
  * @param {Object} [options]
  * @param {string} options.url A string URL of an image file
- * @param {Array} options.coordinates lat, lng coordinates in order clockwise
+ * @param {Array} options.coordinates lng, lat coordinates in order clockwise
  * starting at the top left: tl, tr, br, bl
  * @example
  * var sourceObj = new mapboxgl.ImageSource({
  *    url: 'https://www.mapbox.com/images/foo.png',
  *    coordinates: [
- *        [39.18579907229748, -76.54335737228394],
- *        [39.1838364847587, -76.52803659439087],
- *        [39.17683392507606, -76.5295386314392],
- *        [39.17876344106642, -76.54520273208618]
+ *        [-76.54335737228394, 39.18579907229748],
+ *        [-76.52803659439087, 39.1838364847587],
+ *        [-76.5295386314392, 39.17683392507606],
+ *        [-76.54520273208618, 39.17876344106642]
  *    ]
  * });
  * map.addSource('some id', sourceObj); // add
@@ -67,7 +67,7 @@ ImageSource.prototype = util.inherit(Evented, {
     createTile: function() {
         var map = this.map;
         var coords = this.coordinates.map(function(lnglat) {
-            var loc = LatLng.convert([lnglat[1], lnglat[0]]);
+            var loc = LngLat.convert(lnglat);
             return map.transform.locationCoordinate(loc).zoomTo(0);
         });