+++ test/render.test.js
@@ -51,7 +51,7 @@ function renderTest(style, info, base, key) {
                     add: function() {}
                 }
             },
-            center: info.center || [0, 0],
+            center: info.center ? [info.center[1], info.center[0]] : [0, 0],
             zoom: info.zoom || 0,
             bearing: info.bearing || 0,
             style: style,